refactor: move implementation details to source files and finalize recursion depth constants in validator
